feat: 지출 시각화 부분 통합 및 카테고리 부분 스위치 토글
